# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wormhole 后端服务 API 文档。Keycloak SSO 模式下，前端通过 /auth/sso/login 发起登录，后端写入 HttpOnly Cookie；前端鉴权判断请调用 /users/me 并携带 credentials: include。普通业务接口要求登录，/admin 开头的管理接口额外要求当前用户拥有 admin 角色。",
        "title": "Wormhole Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/admin/announcements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "返回全部公告，可按 status 筛选；草稿和过期公告也会返回。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-announcements"
                ],
                "summary": "管理员获取公告列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "状态：1=发布，0=草稿/下架；不传返回全部",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "公告列表",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "403": {
                        "description": "没有权限",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "500": {
                        "description": "获取公告失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "新增公告；发布时间为空时立即生效，expiresAt 为空时永不过期。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-announcements"
                ],
                "summary": "管理员新增公告",
                "parameters": [
                    {
                        "description": "公告",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAnnouncementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "公告",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "403": {
                        "description": "没有权限",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "500": {
                        "description": "新增公告失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    }
                }
            }
        },
        "/admin/announcements/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "部分更新公告内容、置顶状态、发布时间、到期时间或发布状态。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-announcements"
                ],
                "summary": "管理员编辑公告",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "公告",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAnnouncementRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "公告",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "403": {
                        "description": "没有权限",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "404": {
                        "description": "公告不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "500": {
                        "description": "编辑公告失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    }
                }
            }
        },
        "/admin/announcements/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "更新公告状态，1=发布，0=草稿/下架。",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin-announcements"
                ],
                "summary": "管理员更新公告状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "公告 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "状态",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAnnouncementStatusRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "公告",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "400": {
                        "description": "参数错误",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "403": {
                        "description": "没有权限",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "404": {
                        "description": "公告不存在",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "500": {
                        "description": "更新公告状态失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    }
                }
            }
        },
...
        "/announcements": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    },
                    {
                        "CookieAuth": []
                    }
                ],
                "description": "返回当前登录用户可见的已发布且未过期公告，置顶公告优先。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "announcements"
                ],
                "summary": "获取已发布公告",
                "responses": {
                    "200": {
                        "description": "公告列表",
                        "schema": {
                            "$ref": "#/definitions/response.APIResponse"
                        }
                    },
                    "401": {
                        "description": "未登录或登录已失效",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    },
                    "500": {
                        "description": "获取公告失败",
                        "schema": {
                            "$ref": "#/definitions/dto.ErrorAPIResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CreateAnnouncementRequest": {
            "type": "object",
            "required": [
                "content",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 10000,
                    "example": "本周六 02:00-03:00 进行例行维护。"
                },
                "expiresAt": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "isPinned": {
                    "type": "boolean",
                    "example": false
                },
                "publishedAt": {
                    "type": "string",
                    "example": "2026-07-13T10:00:00Z"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "平台维护通知"
                }
            }
        },
...
        "dto.ErrorAPIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 40101
                },
                "data": {
                    "type": "string",
                    "example": ""
                },
                "error": {
                    "type": "string",
                    "example": "token is expired"
                },
                "message": {
                    "type": "string",
                    "example": "未登录，请先登录"
                },
                "requestId": {
                    "type": "string",
                    "example": "5547e998-1127-4c9d-ae7e-f3508c42b96c"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2026-07-10T16:30:01+08:00"
                }
            }
        },
...
        "dto.UpdateAnnouncementRequest": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string",
                    "maxLength": 10000,
                    "example": "本周六 02:00-03:00 进行例行维护。"
                },
                "expiresAt": {
                    "type": "string",
                    "example": "2026-07-20T10:00:00Z"
                },
                "isPinned": {
                    "type": "boolean",
                    "example": false
                },
                "publishedAt": {
                    "type": "string",
                    "example": "2026-07-13T10:00:00Z"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                },
                "title": {
                    "type": "string",
                    "maxLength": 128,
                    "example": "平台维护通知"
                }
            }
        },
        "dto.UpdateAnnouncementStatusRequest": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
...
        "response.APIResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "error": {},
                "message": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "兼容 Bearer token：Authorization: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "CookieAuth": {
            "description": "SSO 登录成功后由后端写入的 HttpOnly 会话 Cookie，前端 JS 不能读取，fetch/axios 请求需携带 credentials/withCredentials。",
            "type": "apiKey",
            "name": "wormhole_session",
            "in": "cookie"
        }
    }
}